Solo una barra de codigo

+++ src/controllers/tareas.controllers.js
@@ -41,6 +41,11 @@ ctrl.putTarea = async(req, res)=>{
 ctrl.deleteTarea =  async (req,res) =>{
     const id = req.params.id;
 
+    const validacion = await Tareas.findOne({userID:req.user._id, _id:id})
+    if(!validacion){
+        return res.json("No está autorizado");
+    }
+
     const eliminarTarea = await Tareas.deleteOne({_id:id,userID:req.user._id})
 
     res.json("Tarea eliminada" + eliminarTarea)
